sync_es/requirement.txt
python-dotenv
kafka-python
elasticsearch